refactor: 将homepage重构成更方便编辑的格式
